requirements.txt
openai-whisper>=20231117
requests>=2.28.0
python-dotenv>=0.19.0
